+++ tests/test_xavrdebugger.py
@@ -753,3 +753,32 @@ class TestXAvrDebugger(TestCase):
         self.xa.reset()
         self.xa.device.avr.protocol.reset.assert_called_once()
 
+    def test_readmem_sram_masked_register_one_byte(self):
+        self.set_up()
+        self.xa._iooffset.return_value = 0
+        self.xa._masked_registers = [0x35, 0x21, 0x26]
+        self.assertEqual(self.xa.sram_masked_register(0x21, 1), bytearray([0x00]))
+
+    def test_readmem_sram_masked_register_bytearray(self):
+        self.set_up()
+        self.xa._iooffset.return_value = 0
+        self.xa._masked_registers = [0x35, 0x21, 0x26]
+        sram = list(reversed(range(0x35)))
+        def access_sram(ix, length):
+            return bytearray(sram[ix:ix+length])
+        self.xa.sram_read = MagicMock(side_effect=access_sram)
+        self.assertEqual(self.xa.sram_masked_register(0x25, 3), bytearray([0x0F, 0x00, 0x0D]))
+        self.assertEqual(self.xa.sram_masked_register(0x24, 3), bytearray([0x10, 0x0F, 0x00]))
+
+    def test_writemem_sram_ronly_register_one_byte(self):
+        self.set_up()
+        self.xa._iooffset.return_value = 0
+        self.xa._ronly_registers = [15, 1, 6]
+        self.assertEqual(self.xa.sram_masked_write(1, bytearray([0x55])), "OK")
+
+    def test_writemem_sram_ronly_register_bytearray(self):
+        self.set_up()
+        self.xa.get_iooffset.return_value = 0
+        self.xa._ronly_registers = [15, 1, 6, 0]
+        self.assertEqual(self.xa.sram_masked_write(1, bytearray([0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07])), "OK")
+